+++ Guardar/GuardarJavas/DadosPagamentoFatura.java
@@ -3,5 +3,6 @@ package br.desafio.prodiga.dto.Fatura;
 import java.time.LocalDate;
 
 public record DadosPagamentoFatura(
+    Long clienteId,
     LocalDate dataPagamento
 ) {}
