+++ desafio-casdastro/src/main/java/br/desafio/prodiga/Model/DadosEndereco.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.dto;
+package br.desafio.prodiga.Model;
 
 public record DadosEndereco(
     String logradouro,
