+++ desafio-casdastro/src/main/java/br/desafio/prodiga/Model/Fatura.java
@@ -5,47 +5,47 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Random;
 
-import br.desafio.prodiga.dto.Fatura.DataFaturas;
+import br.desafio.prodiga.dto.Fatura.DadosCadastroFatura;
 import br.desafio.prodiga.dto.Fatura.Situacao;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
+
+import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+
 
 @Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@RequiredArgsConstructor
+@Data
 @EqualsAndHashCode(of = "id")
 @Table(name = "faturas")
+
 public class Fatura implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String numfatura;
-
+    private String numerofatura;
+   
     private int mes;
 
+    
     private int ano;
 
+    
     private Double valor;
 
+    
     private LocalDate dataVencimento;
 
-    // @Column(nullable = false, unique = true)
+    
     private String codigoBoleto;
 
     @Enumerated(EnumType.STRING)
@@ -55,35 +55,35 @@ public class Fatura implements Serializable {
 
     private LocalDateTime dataGeracao = LocalDateTime.now();
 
-    @ManyToOne
-    @JoinColumn(name = "cliente_id")
+    @ManyToOne(fetch = FetchType.LAZY)
     private Cliente cliente;
+     
+    public void gerarNumFatura(Random random) {
+        this.numerofatura = String.format("FAT-%d%02d-%06d", 
+            LocalDateTime.now().getYear(),
+            LocalDateTime.now().getMonthValue(),
+            random.nextInt(10000));
+    }
+    public void gerarCodigoBoleto(Random random){
+        this.codigoBoleto = String.format("FAT-%d%02d-%06d",
+        LocalDateTime.now().getYear(), LocalDateTime.now().getDayOfMonth(),
+        random.nextInt(10000));
+    }
 
-    // Construtor personalizado com foco nas datas das faturas
-    public Fatura(DataFaturas dados) {
+     public Fatura(DadosCadastroFatura dados){
+        this.id = dados.id();
         this.ano = dados.ano();
         this.mes = dados.mes();
-        this.dataGeracao = dados.dataGeracao();
-        this.dataPagamento = dados.dataPagamento();
+        this.valor = dados.valor();
         this.dataVencimento = dados.dataVencimento();
-    }
+        this.codigoBoleto = dados.codigoBoleto();
+        this.situacao = dados.situacao();
+        this.dataPagamento = dados.dataPagamento();
+        this.dataGeracao = dados.dataGeracao();
+        this.cliente = dados.cliente();
+     }
 
-    
 
-    public void gerarNumFatura() {
-        this.numfatura = "FAT-" + LocalDateTime.now().getYear() +
-                LocalDateTime.now().getMonthValue() +
-                "-" + String.format("%06d", new Random().nextInt(10000));
-    }
-    public void setDatapPagamento(LocalDate dataPagamento) {
-        this.dataPagamento = dataPagamento.atStartOfDay();
-    }
-
-    public void setAnoReferencia(int ano) {
-        this.ano = ano;
-    }
-
-    public void setMesReferenia(int mes) {
-        this.mes = mes;
-    }
+     
+     
 }
