+++ desafio-casdastro/src/main/java/br/desafio/prodiga/Repository/FaturaRepository.java
@@ -2,13 +2,17 @@ package br.desafio.prodiga.Repository;
 
 import java.util.List;
 
+import org.springframework.boot.autoconfigure.data.web.SpringDataWebProperties.Pageable;
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import br.desafio.prodiga.Model.Fatura;
+import br.desafio.prodiga.dto.Fatura.ListaFatura;
 import br.desafio.prodiga.dto.Fatura.Situacao;
 @Repository
 public interface FaturaRepository extends JpaRepository<Fatura, Long> {
        List<Fatura> findBySituacao(Situacao situacao);
        List<Fatura> findByClienteId(Long clienteId);
+       Page<ListaFatura> findByLista(Pageable paginacao);
 }
