+++ desafio-casdastro/src/main/java/br/desafio/prodiga/dto/Cliente/DadosClientes.java
@@ -1,6 +1,6 @@
 package br.desafio.prodiga.dto.Cliente;
 
-import br.desafio.prodiga.dto.DadosEndereco;
+import br.desafio.prodiga.Model.DadosEndereco;
 
 public record DadosClientes(
     String nome,
