--- desafio-casdastro/src/test/java/br/desafio/prodiga/Service/ClienteServicoTest.java
@@ -1,22 +1,0 @@
-package br.desafio.prodiga.Service;
-
-import org.junit.jupiter.api.Test;
-
-public class ClienteServicoTest {
-    @Test
-    void testCadastrar() {
-
-    }
-
-    @Test
-    void testApagarClientes() {
-        
-    }
-
-    @Test
-    void testCadastrar2() {
-        
-    }
-
- 
-}
