--- desafio-casdastro/src/test/java/br/desafio/prodiga/Service/FaturaServiceTest.java
@@ -1,10 +1,0 @@
-package br.desafio.prodiga.Service;
-
-import org.junit.jupiter.api.Test;
-
-public class FaturaServiceTest {
-    @Test
-    void testGerarFaturas() {
-   
-    }
-}
